refactor: TS in  burger-ingredients
